fix: extract refdiff run settings as component
