+++ config-ui/src/components/pipelines/ProviderSettings.jsx
@@ -6,20 +6,21 @@ import {
   Button,
   ButtonGroup,
   FormGroup,
-  Popover,
+  // Popover,
   InputGroup,
   MenuItem,
-  Menu,
+  // Menu,
   Intent,
   TagInput,
-  Checkbox,
-  Tag,
-  Switch,
+  // Checkbox,
+  // Tag,
+  // Switch,
   Tooltip,
-  Colors,
+  // Colors,
 } from '@blueprintjs/core'
 import { Select } from '@blueprintjs/select'
-import RefDiffTasksMenu from '@/components/menus/RefDiffTasksMenu'
+// import RefDiffTasksMenu from '@/components/menus/RefDiffTasksMenu'
+import RefDiffSettings from '@/components/pipelines/pipeline-settings/refdiff'
 
 const ProviderSettings = (props) => {
   const {
@@ -54,33 +55,33 @@ const ProviderSettings = (props) => {
 
   let providerSettings = null
 
-  const [refDiffOldTag, setRefDiffOldTag] = useState('')
-  const [refDiffNewTag, setRefDiffNewTag] = useState('')
+  // const [refDiffOldTag, setRefDiffOldTag] = useState('')
+  // const [refDiffNewTag, setRefDiffNewTag] = useState('')
 
-  const handleRefDiffTaskSelect = (e, task) => {
-    setRefDiffTasks(t => t.includes(task.task)
-      ? t.filter(t2 => t2 !== task.task)
-      : [...new Set([...t, task.task])])
-  }
+  // const handleRefDiffTaskSelect = (e, task) => {
+  //   setRefDiffTasks(t => t.includes(task.task)
+  //     ? t.filter(t2 => t2 !== task.task)
+  //     : [...new Set([...t, task.task])])
+  // }
 
-  const createRefDiffPair = (oldRef, newRef) => {
-    return {
-      oldRef,
-      newRef
-    }
-  }
+  // const createRefDiffPair = (oldRef, newRef) => {
+  //   return {
+  //     oldRef,
+  //     newRef
+  //   }
+  // }
 
-  const addRefDiffPairObject = (oldRef, newRef) => {
-    setRefDiffPairs(pairs => (!pairs.some(p => p.oldRef === oldRef && p.newRef === newRef))
-      ? [...pairs, { oldRef, newRef }]
-      : [...pairs])
-    setRefDiffNewTag('')
-    setRefDiffOldTag('')
-  }
+  // const addRefDiffPairObject = (oldRef, newRef) => {
+  //   setRefDiffPairs(pairs => (!pairs.some(p => p.oldRef === oldRef && p.newRef === newRef))
+  //     ? [...pairs, { oldRef, newRef }]
+  //     : [...pairs])
+  //   setRefDiffNewTag('')
+  //   setRefDiffOldTag('')
+  // }
 
-  const removeRefDiffPairObject = (oldRef, newRef) => {
-    setRefDiffPairs(pairs => pairs.filter(p => !(p.oldRef === oldRef && p.newRef === newRef)))
-  }
+  // const removeRefDiffPairObject = (oldRef, newRef) => {
+  //   setRefDiffPairs(pairs => pairs.filter(p => !(p.oldRef === oldRef && p.newRef === newRef)))
+  // }
 
   // useEffect(() => {
   //   console.log('>>>> REF DIFF PAIRS ARRAY...', refDiffPairs)
@@ -368,211 +369,21 @@ const ProviderSettings = (props) => {
       break
     case Providers.REFDIFF:
       providerSettings = (
-        <>
-          <FormGroup
-            disabled={isRunning || !isEnabled(providerId)}
-            label={<strong>Repository ID<span className='requiredStar'>*</span></strong>}
-            labelInfo={<span style={{ display: 'block' }}>Enter Repo Column ID</span>}
-            inline={false}
-            labelFor='refdiff-repo-id'
-            className=''
-            contentClassName=''
-            style={{ minWidth: '280px', marginBottom: 'auto' }}
-            fill
-          >
-            <InputGroup
-              id='refdiff-repo-id'
-              disabled={isRunning || !isEnabled(providerId)}
-              placeholder='eg. github:GithubRepo:384111310'
-              value={refDiffRepoId}
-              onChange={(e) => setRefDiffRepoId(e.target.value)}
-              className='input-refdiff-repo-id'
-              autoComplete='off'
-              fill={false}
-              style={{ marginBottom: '10px' }}
-            />
-          </FormGroup>
-          <FormGroup
-            disabled={isRunning || !isEnabled(providerId)}
-            label={(
-              <strong>Tasks<span className='requiredStar'>*</span>
-                <span
-                  className='badge-count'
-                  style={{
-                    opacity: isEnabled(providerId) ? 0.5 : 0.1
-                  }}
-                >{refDiffTasks.length}
-                </span>
-              </strong>
-              )}
-            labelInfo={<span style={{ display: 'block' }}>Select Tasks to Execute</span>}
-            inline={false}
-            labelFor='refdiff-tasks'
-            className=''
-            contentClassName=''
-            style={{ marginLeft: '12px', marginRight: '12px', marginBottom: 'auto' }}
-            fill
-          >
-
-            <Popover
-              className='provider-tasks-popover'
-              disabled={isRunning || !isEnabled(providerId)}
-              content={(
-                <RefDiffTasksMenu onSelect={handleRefDiffTaskSelect} selected={refDiffTasks} />
-              )}
-              placement='top-center'
-            >
-              <ButtonGroup disabled={isRunning || !isEnabled(providerId)}>
-                <Button
-                  disabled={isRunning || !isEnabled(providerId)}
-                  icon='menu'
-                  // text='Choose Tasks'
-                  text={refDiffTasks.length > 0
-                    ? <>Choose Tasks <Tag intent={Intent.PRIMARY} round>{refDiffTasks.length}</Tag></>
-                    : 'Choose Tasks'}
-                />
-                <Button
-                  icon='eraser'
-                  // text={refDiffTasks.length > 0
-                  //   ? <><Tag intent={Intent.PRIMARY} round>{refDiffTasks.length}</Tag></>
-                  //   : ''}
-                  intent={Intent.WARNING}
-                  disabled={isRunning || !isEnabled(providerId) || refDiffTasks.length === 0}
-                  onClick={() => setRefDiffTasks([])}
-                />
-              </ButtonGroup>
-            </Popover>
-          </FormGroup>
-          <div style={{ display: 'flex', flex: 1, width: '100%' }}>
-            <FormGroup
-              disabled={isRunning || !isEnabled(providerId)}
-              label={(
-                <strong>Tags<span className='requiredStar'>*</span>
-                  <span
-                    className='badge-count'
-                    style={{
-                      opacity: isEnabled(providerId) ? 0.5 : 0.1
-                    }}
-                  >{refDiffPairs.length}
-                  </span>
-                </strong>
-              )}
-              labelInfo={<span style={{ display: 'block' }}>Specify tag Ref Pairs</span>}
-              inline={false}
-              labelFor='refdiff-pair-newref'
-              className=''
-              contentClassName=''
-              style={{ minWidth: '588px', marginBottom: 'auto' }}
-              fill={false}
-            >
-              <div style={{ display: 'flex' }}>
-                <div>
-                  <InputGroup
-                    id='refdiff-pair-newref'
-                    // round='true'
-                    leftElement={(
-                      <Tag
-                        intent={Intent.WARNING} style={{
-                          opacity: isEnabled(providerId) ? 1 : 0.3
-                        }}
-                      >New Ref
-                      </Tag>
-                    )}
-                    inline={true}
-                    disabled={isRunning || !isEnabled(providerId)}
-                    placeholder='eg. refs/tags/v0.6.0'
-                    value={refDiffNewTag}
-                    onChange={(e) => setRefDiffNewTag(e.target.value)}
-                    autoComplete='off'
-                    fill={false}
-                    // small
-                  />
-                </div>
-                <div style={{ marginLeft: '10px', marginRight: '10px' }}>
-                  {/* <label>Old Ref</label> */}
-                  <InputGroup
-                    id='refdiff-pair-oldref'
-                    // round='true'
-                    leftElement={(
-                      <Tag
-                        style={{
-                          opacity: isEnabled(providerId) ? 1 : 0.3
-                        }}
-                      >Old Ref
-                      </Tag>
-                    )}
-                    inline={true}
-                    disabled={isRunning || !isEnabled(providerId)}
-                    placeholder='eg. refs/tags/v0.5.0'
-                    value={refDiffOldTag}
-                    onChange={(e) => setRefDiffOldTag(e.target.value)}
-                    autoComplete='off'
-                    fill={false}
-                    rightElement={(
-                      <Tooltip content='Add Tag Pair'>
-                        <Button
-                          className='btn-add-tagpair'
-                          intent={Intent.PRIMARY}
-                          disabled={!refDiffOldTag || !refDiffNewTag || refDiffOldTag === refDiffNewTag}
-                          icon='plus'
-                          small
-                          style={{}}
-                          onClick={() => addRefDiffPairObject(refDiffOldTag, refDiffNewTag)}
-                        />
-                      </Tooltip>
-
-                    )}
-                    // small
-                  />
-                </div>
-                <div style={{ display: 'flex', justifyContent: 'flex-end' }}>
-                  {/* <Button icon='remove' minimal small style={{ marginTop: 'auto', alignSelf: 'center' }} /> */}
-                  {/* <Button
-                    intent={Intent.PRIMARY}
-                    disabled={!refDiffOldTag || !refDiffNewTag || refDiffOldTag === refDiffNewTag}
-                    icon='add'
-                    small
-                    style={{ marginTop: 'auto', alignSelf: 'center' }}
-                    onClick={() => addRefDiffPairObject(refDiffOldTag, refDiffNewTag)}
-                  /> */}
-                </div>
-              </div>
-              <div
-                style={{
-                  borderRadius: '4px',
-                  padding: '10px',
-                  margin: '5px 0'
-                }}
-              >
-                {refDiffPairs.map((pair, pairIdx) => (
-                  <div key={`refdiff-added-pairs-itemkey-$${pairIdx}`} style={{ display: 'flex' }}>
-                    <div style={{ flex: 1 }}>
-                      <Tag intent={Intent.WARNING} round='false' small>new</Tag> {pair.newRef}
-                    </div>
-                    <div style={{ flex: 1, marginLeft: '10px', marginRight: '10px' }}>
-                      <Tag round='false' small>old</Tag> {pair.oldRef}
-                    </div>
-                    <div style={{ display: 'flex', justifyContent: 'flex-end' }}>
-                      <Button
-                        className='btn-remove-tagpair'
-                        icon='remove'
-                        minimal
-                        small
-                        style={{ marginTop: 'auto', alignSelf: 'center' }}
-                        onClick={() => removeRefDiffPairObject(pair.oldRef, pair.newRef)}
-                      />
-                    </div>
-                  </div>
-                ))}
-                {refDiffPairs.length === 0 && <><span style={{ color: Colors.GRAY3 }}>( No Tag Pairs Added)</span></>}
-              </div>
-            </FormGroup>
-          </div>
-        </>
-
+        <RefDiffSettings
+          providerId={providerId}
+          repoId={refDiffRepoId}
+          tasks={refDiffTasks}
+          pairs={refDiffPairs}
+          setRepoId={setRefDiffRepoId}
+          setTasks={setRefDiffTasks}
+          setPairs={setRefDiffPairs}
+          isRunning={isRunning}
+          isEnabled={isEnabled}
+        />
       )
       break
     default:
+      providerSettings = null
       break
   }
